application/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "api"
authors = ["shun198"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.24.2"
psycopg2 = "^2.9.6"
gunicorn = "^20.1.0"
drf-nested-routers = "^0.93.4"
celery = "^5.2.7"
django-celery-beat = "^2.4.0"
redis = "^4.3.4"
django-cors-headers = "^3.13.0"
Authlib = "^1.1.0"
django-filter = "^22.1"
boto3 = "^1.26.22"
injector = "^0.20.1"
django-ses = "^3.2.2"
django-storages = "^1.13.1"
pydantic = "^1.10.6"
toml = "^0.10.2"
pillow = "^9.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
pytest-django = "^4.5.2"
pytest-sugar = "^0.9.6"
pytest-xdist = "^3.0.2"
django-debug-toolbar = "^3.8.1"
pytest-bdd = "^6.1.1"
pytest-clarity = "^1.0.1"
pytest-custom-exit-code = "^0.3.0"
factory-boy = "^3.2.1"
pytest-mock = "^3.10.0"
black = "^22.10.0"
isort = "^5.11.4"
django-extensions = "^3.2.1"
pdoc = "^13.0.0"
pytest-github-actions-annotate-failures = "^0.2.0"

[tool.isort]
profile = "black"

[tool.black]
line-length = 79
include = '\.py$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-v -s --durations=0"

[tool.pytest]
DJANGO_SETTINGS_MODULE = "project.settings.local"

# logの設定（ローカル環境）
[local.logging]
version = 1
[local.logging.formatters.simple]
format = "[%(levelname)s] %(name)s %(asctime)s - %(message)s "
[local.logging.handlers.consoleHandler]
class = "logging.StreamHandler"
level = "DEBUG"
formatter = "simple"
stream = "ext://sys.stdout"
[local.logging.handlers.applicationHandler]
class = "logging.handlers.TimedRotatingFileHandler"
when = "D"
level = "INFO"
formatter = "simple"
filename = "./output/application.log"
[local.logging.handlers.errorHandler]
class = "logging.handlers.TimedRotatingFileHandler"
when = "D"
level = "ERROR"
formatter = "simple"
filename = "./output/emergency.log"
[local.logging.loggers.console]
level = "DEBUG"
handlers = ["consoleHandler"]
propagate = "no"
[local.logging.loggers.application]
level = "DEBUG"
handlers = ["consoleHandler","applicationHandler"]
propagate = "no"
[local.logging.loggers.emergency]
level = "DEBUG"
handlers = ["consoleHandler","errorHandler"]
propagate = "no"

# logの設定（dev環境）
[dev.logging]
version = 1
[dev.logging.formatters.simple]
format = "[%(levelname)s] %(name)s - %(message)s "
[dev.logging.handlers.consoleHandler]
class = "logging.StreamHandler"
level = "DEBUG"
formatter = "simple"
stream = "ext://sys.stdout"
[dev.logging.loggers.console]
level = "DEBUG"
handlers = ["consoleHandler"]
propagate = "no"
[dev.logging.loggers.application]
level = "DEBUG"
handlers = ["consoleHandler"]
propagate = "no"
[dev.logging.loggers.emergency]
level = "DEBUG"
handlers = ["consoleHandler"]
propagate = "no"
